Added separation of analysis plugins + updated analysis docs
